Refactor queue-song route to handle errors and improve code readability

+++ src/app/api/queue-song/get-next/[shopId]/route.ts
@@ -1,35 +1,33 @@
-import { NextResponse, NextRequest } from "next/server";
-import prisma from "@/lib/prisma";
+import { NextResponse, NextRequest } from 'next/server'
+import prisma from '@/lib/prisma'
 
-export async function GET(
-  request: NextRequest,
-  { params: { shopId } }: { params: { shopId: string } }
-) {
+export async function GET(request: NextRequest, { params: { shopId } }: { params: { shopId: string } }) {
   try {
     // get the first song in the queue
+    const searchParams = request.nextUrl.searchParams
+    const method = searchParams.get('method')
+
     const song = await prisma.queueSong.findFirst({
       where: {
         coffeeShopId: shopId,
       },
       orderBy: {
-        id: "asc",
+        id: 'asc',
       },
-    });
+    })
 
     if (!song) {
-      return NextResponse.json(
-        { message: "No songs in the queue" },
-        { status: 404 }
-      );
+      return NextResponse.json({ message: 'No songs in the queue' }, { status: 404 })
     }
 
     // remove the song from the queue
-    await prisma.queueSong.delete({
-      where: {
-        id: song.id,
-      },
-    });
-
+    if (method === 'remove') {
+      await prisma.queueSong.delete({
+        where: {
+          id: song.id,
+        },
+      })
+    }
     // get the updated queue
     const queue = await prisma.queueSong.findMany({
       where: {
@@ -38,17 +36,14 @@ export async function GET(
       include: {
         song: true,
       },
-    });
+    })
 
     // return the next 2 songs in the queue
-    const nextSongs = queue.slice(0, 2);
+    const nextSongs = queue.slice(0, 2)
 
-    return NextResponse.json(nextSongs);
+    return NextResponse.json(nextSongs)
   } catch (error) {
-    console.error("Error fetching or processing the queue:", error);
-    return NextResponse.json(
-      { message: "An error occurred while processing the queue" },
-      { status: 500 }
-    );
+    console.error('Error fetching or processing the queue:', error)
+    return NextResponse.json({ message: 'An error occurred while processing the queue' }, { status: 500 })
   }
 }
